+++ CHANGELOG.md
@@ -6,7 +6,10 @@
 
 - Remove `component.BaseProcessorFactory`, use `processorhelper.NewFactory` instead (#4175)
 - Move `service/parserprovider` package to `config/configmapprovider` (#4206)
-- Rename `pdata.AnyValueArray` to `AttributeSlice` (#4325)
+- Renamed slice-valued `pdata` types and functions for consistency. (#4325)
+  - Rename `pdata.AnyValueArray` to `AttributeSlice`
+  - Rename `ArrayVal()` to `SliceVal()`
+  - Rename `SetArrayVal()` to `SetSliceVal()`
 
 ## v0.38.0 Beta